+++ Model/BodyWeightEntryDot.cs
@@ -3,10 +3,10 @@ namespace WorkoutLog.Model;
 public class BodyWeightEntryDot : BodyWeightEntryPropertyChanged
 {
     public DateTime date { get; set; }
-    public int weight { get; set; }
+    public double weight { get; set; }
     public int y_adjustment { get; set; }
-    public int highest_value { get; set; }
-    public int lowest_value { get; set; }
+    public double highest_value { get; set; }
+    public double lowest_value { get; set; }
 
     private bool _isCurrentDateSelected;
     public bool is_current_date_selected
