+++ Popup/BodyWeightAddPopup.xaml.cs
@@ -21,7 +21,7 @@ public partial class BodyWeightAddPopup
         if (weight_update_string != null && weight_update_string.Length != 0) /* if weight field is not empty */
         {
             weight_update_string = weight_update_string.ToString();
-            int weight_update = int.Parse(weight_update_string);
+            double weight_update = double.Parse(weight_update_string);
 
             DateTime date = record_date.Date;
